Controller 중복 수정
