+++ src/main/java/com/goormitrip/goormitrip/repository/UserRepository.java
@@ -3,5 +3,9 @@ package com.goormitrip.goormitrip.repository;
 import com.goormitrip.goormitrip.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
 }
