+++ src/main/java/com/goormitrip/goormitrip/service/UserService.java
@@ -1,0 +1,29 @@
+package com.goormitrip.goormitrip.service;
+
+import com.goormitrip.goormitrip.dto.SignupRequestDto;
+import com.goormitrip.goormitrip.repository.UserRepository;
+import lombok.AllArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class UserService implements UserDetailsService {
+
+    private UserRepository userRepository;
+    private PasswordEncoder passwordEncoder;
+
+    public void processSignup(SignupRequestDto signupRequestDto) {
+        userRepository.save(signupRequestDto.toUser(passwordEncoder));
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findByEmail(username)
+                .orElseThrow(() ->
+                        new UsernameNotFoundException("username '" + username + "' not found"));
+    }
+}
